nopCommerce 4.n - IMPORTANT RISK - When a Vendor is DISABLED or DELETED - the associated products remain available to the customer to order, and this can be embarrassing as it was last week for us.
I have created 2 SQL Triggers on the Vendor Table that are for :-

Vendor UPDATEd:

If the Vendor status ACTIVE is set to 0 the all associated Products are set to PUBLISHED = 0, and DISABLE BUY-BUTTON = 1

and vice versa. (This must be handled carefully are there may be products that were originally unpublished, so perhaps apply further rules of your own to ensure this does not affect you. For us this is not a problem.

Vendor DELETEd:

If the Vendor if DELETED, then when the Vendor Deleted Action occurs, all associated Products are set to PUBLISHED = 0, and DELETED = 1
I thought about catering fir a scenario where the Products could be reinstated if the Vendor was recreated but, there are too many variables to different situations. My scripts cans serve as a basis for anyone needing to expand on this.

I have use the SQL approach and handled it from the back end from a performance point of view. It is entirely up to you if you wish to convert this to a function and handled the cases from the application layer. Whatever floats you boat.

NB! Just be aware if this underlying risk though of the Products remaining available even after the Vendor is Deleted or Inactive!